fix(AskConsent): parse dialog to its own component and check for messages to display the dialog only when needed
